+++ CHANGELOG.md
@@ -6,13 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
-
 * Add `deleteFiles` option and `--delete-files` flag to delete all
   files/directories matching some glob patterns after conversion.
 * If a package already has a `package.json`, it will be merged with the newly
   generated one.
 * Added `--flat` and `--private` CLI flags to control those options in the new
   `package.json`. Both default to `false`.
+* `deleteFiles` option will no longer delete any file from `node_modules/` or
+  `bower_components/`.
 <!-- Add new, unreleased changes here. -->
 
 ## [0.3.0] - 2017-11-28